test_individual_chunks.py
#!/usr/bin/env python3
"""
Test Individual Chunk Results

Verifies that chunks are returned as individual results, not reconstructed.
"""

from PIL import Image
import numpy as np
from image_preprocessing_pipeline import preprocess_images_batch

def test_individual_chunk_results():
    """Test that chunks are returned as individual results"""
    
    print("=" * 60)
    print("Testing Individual Chunk Results")
    print("=" * 60)
    
    # Create a test image
    print("\n1. Creating test image (101708 x 1229)...")
    test_array = np.random.randint(0, 255, (1229, 101708, 3), dtype=np.uint8)
    test_image = Image.fromarray(test_array)
    
    # Preprocess directly with PIL image
    print("\n2. Preprocessing with resize and chunk...")
    from image_preprocessing_pipeline import ImagePreprocessor
    
    preprocessor = ImagePreprocessor(
        target_width=31778,
        target_height=384,
        chunk_width=2048,
        chunk_height=384
    )
    
    resized_image, chunks = preprocessor.process_image(test_image)
    chunk_info = preprocessor.get_chunk_info(chunks)
    
    # Create preprocessed data structure
    preprocessed = [{
        'original_filename': 'test_image.png',
        'original_size': test_image.size,
        'resized_image': resized_image,
        'chunks': chunks,
        'chunk_info': chunk_info
    }]
    
    print(f"   ✅ Preprocessed 1 image")
    print(f"   ✅ Original size: {preprocessed[0]['original_size']}")
    print(f"   ✅ Resized to: {preprocessed[0]['resized_image'].size}")
    print(f"   ✅ Total chunks: {preprocessed[0]['chunk_info']['total_chunks']}")
    
    # Verify chunk structure
    print("\n3. Verifying individual chunks...")
    chunks = preprocessed[0]['chunks']
    
    print(f"   Total chunks: {len(chunks)}")
    
    for i, chunk in enumerate(chunks[:3]):  # Show first 3
        print(f"   Chunk {i+1}:")
        print(f"      - Image size: {chunk['chunk'].size}")
        print(f"      - X offset: {chunk['x_offset']}")
        print(f"      - Width: {chunk['width']}")
        print(f"      - Height: {chunk['height']}")
    
    # Simulate what detection results should look like
    print("\n4. Simulating detection results format...")
    print("   Each chunk should become an individual result:")
    
    for i, chunk in enumerate(chunks[:3]):  # Show first 3
        result = {
            'filename': f"test_image.png_chunk_{i+1}",
            'original_filename': 'test_image.png',
            'image': chunk['chunk'],  # Individual chunk image
            'matches': [],  # Would contain detections
            'chunk_index': i,
            'chunk_info': {
                'chunk_number': i + 1,
                'total_chunks': len(chunks),
                'x_offset': chunk['x_offset'],
                'width': chunk['width'],
                'height': chunk['height']
            }
        }
        print(f"\n   Result {i+1}:")
        print(f"      - Filename: {result['filename']}")
        print(f"      - Image size: {result['image'].size}")
        print(f"      - Chunk {result['chunk_info']['chunk_number']}/{result['chunk_info']['total_chunks']}")
    
    print("\n" + "=" * 60)
    print("✅ INDIVIDUAL CHUNK RESULTS VERIFIED!")
    print("=" * 60)
    print("\nExpected behavior:")
    print(f"✅ {len(chunks)} individual results (one per chunk)")
    print("✅ Each result has its own chunk image")
    print("✅ Each result shows chunk number and offset")
    print("✅ Results Analysis tab will display each chunk separately")
    
    return True

if __name__ == "__main__":
    try:
        test_individual_chunk_results()
        print("\n🎉 TEST PASSED!")
    except Exception as e:
        print(f"\n❌ TEST FAILED: {str(e)}")
        import traceback
        traceback.print_exc()
